requirements.txt
mysql-connector-python
flask
flask_socketio
pyzt